refactor: Move built-in section definitions to elf.rs
