+++ libwild/src/lib.rs
@@ -253,7 +253,8 @@ impl Linker {
 
         let mut output = file_writer::Output::new(args, output_kind);
 
-        let mut output_sections = OutputSections::with_base_address(output_kind.base_address());
+        let mut output_sections =
+            OutputSections::with_base_address::<elf::File>(output_kind.base_address());
 
         let mut layout_rules_builder = LayoutRulesBuilder::default();
